+++ PolygonChecker/input.h
@@ -1,5 +1,5 @@
 #pragma once
 //Maksymp-input for rectangle-interface
-void getRectangleCoordinates(float** rectangle_coordinates);
+void getRectangleCoordinates(double** rectangle_coordinates);
 int inputIsSuitable(int symbols_returned);
